Render service pages as client components
